Protecting user notes with auth middleware

+++ backend/src/controllers/NoteController.ts
@@ -2,6 +2,7 @@ import { RequestHandler } from "express";
 import NoteModel from "../models/Note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 interface CreateNodeBody {
     title?: string,
@@ -20,8 +21,12 @@ interface UpdateNoteBody {
 class ApiController {
     // [GET] /api/notes
     getNotes: RequestHandler = async (req, res, next) => {
+        const authenticatedUserId = req.session.userId;
+
         try {
-            const notes = await NoteModel.find();
+            assertIsDefined(authenticatedUserId);
+
+            const notes = await NoteModel.find({ userId: authenticatedUserId });
             res.status(200).json(notes);
         } catch (error) {
             next(error);
@@ -31,7 +36,11 @@ class ApiController {
     // [GET] /api/notes/:noteId
     getNote: RequestHandler = async (req, res, next) => {
         const noteId = req.params.noteId;
+        const authenticatedUserId = req.session.userId;
+
         try {
+            assertIsDefined(authenticatedUserId);
+
             if (!mongoose.isValidObjectId(noteId)) {
                 throw createHttpError(400, "Invalid note id");
             }
@@ -41,6 +50,11 @@ class ApiController {
             if (!note) {
                 throw createHttpError(404, "Note not found");
             }
+
+            if (!note.userId?.equals(authenticatedUserId)) {
+                throw createHttpError(401, "You cannot access this note");
+            }
+
             res.status(200).json(note);
         } catch (error) {
             next(error);
@@ -51,11 +65,16 @@ class ApiController {
     createNote: RequestHandler<unknown, unknown, CreateNodeBody, unknown> = async (req, res, next) => {
         const title = req.body.title;
         const text = req.body.text;
+        const authenticatedUserId = req.session.userId;
+
         try {
+            assertIsDefined(authenticatedUserId);
+
             if (!title) {
                 throw createHttpError(400, "Note must have a title");
             }
-            const newNote = await NoteModel.create({ title, text });
+
+            const newNote = await NoteModel.create({ userId: authenticatedUserId, title, text });
             res.status(200).json(newNote);
         } catch (error) {
             next(error);
@@ -65,9 +84,12 @@ class ApiController {
     // [PATCH] /api/notes/:noteId
     updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBody, unknown> = async (req, res, next) => {
         const noteId = req.params.noteId;
-        const newTitle = req.body.title;
-        const newText = req.body.text;
+        const { title: newTitle, text: newText } = req.body;
+        const authenticatedUserId = req.session.userId;
+
         try {
+            assertIsDefined(authenticatedUserId);
+
             if (!mongoose.isValidObjectId(noteId)) {
                 throw createHttpError(400, "Invalid note id");
             }
@@ -80,6 +102,10 @@ class ApiController {
 
             if (!note) {
                 throw createHttpError(404, "Not note found");
+            }
+
+            if (!note.userId?.equals(authenticatedUserId)) {
+                throw createHttpError(401, "You cannot access this note");
             }
 
             note.title = newTitle;
@@ -96,7 +122,11 @@ class ApiController {
     // [DELETE] /api/notes/:noteId
     deleteNote: RequestHandler = async (req, res, next) => {
         const noteId = req.params.noteId;
+        const authenticatedUserId = req.session.userId;
+
         try {
+            assertIsDefined(authenticatedUserId);
+
             if (!mongoose.isValidObjectId(noteId)) {
                 throw createHttpError(400, "Invalid note id");
             }
@@ -105,6 +135,10 @@ class ApiController {
 
             if (!note) {
                 throw createHttpError(404, "Note not found");
+            }
+
+            if (!note.userId?.equals(authenticatedUserId)) {
+                throw createHttpError(401, "You cannot access this note");
             }
 
             await note.deleteOne();
